Lesson13 (#15)
* lesson 13 JPA SQL

* lesson 13 JPA SQL update long to Long

* lesson 13 JPA SQL update gradle

* gradle fix

* lesson 13 flyway fix

* lesson 13 tests fix

* migration fix

---------

+++ src/test/groovy/ie/futurecollars/invoicing/service/InvoiceServiceIntegrationTest.groovy
@@ -27,9 +27,9 @@ class InvoiceServiceIntegrationTest extends Specification {
 
         then:
         ids == (1..invoices.size()).collect()
-        ids.forEach({ assert service.getById(it).isPresent() })
-        ids.forEach({ assert service.getById(it).get().getId() == it })
-        ids.forEach({ assert service.getById(it).get() == invoices.get(it - 1) })
+        ids.forEach{ assert service.getById(it).isPresent() }
+        ids.forEach{ assert service.getById(it).get().getId() == it }
+        ids.forEach{ assert service.getById(it).get() == invoices.get((int)it - 1) }
     }
 
     def "get by id returns empty optional when there is no invoice with given id"() {
@@ -48,15 +48,15 @@ class InvoiceServiceIntegrationTest extends Specification {
 
         expect:
         service.getAll().size() == invoices.size()
-        service.getAll().forEach({ assert it == invoices.get(it.getId() - 1) })
+        service.getAll().forEach{ assert it == invoices.get((int) it.getId() - 1) }
 
         when:
         service.delete(1)
 
         then:
         service.getAll().size() == invoices.size() - 1
-        service.getAll().forEach({ assert it == invoices.get(it.getId() - 1) })
-        service.getAll().forEach({ assert it.getId() != 1 })
+        service.getAll().forEach{ assert it == invoices.get((int) it.getId() - 1) }
+        service.getAll().forEach{ assert it.getId() != 1 }
     }
 
     def "can delete all invoices"() {
